requirements.txt
aioredis==1.2.0
amqp==2.4.2
asgiref==3.1.2
async-timeout==3.0.1
attrs==19.1.0
autobahn==19.3.3
Automat==0.7.0
billiard==3.6.0.0
boto3==1.9.130
botocore==1.12.130
celery==4.3.0
certifi==2019.3.9
channels==2.2.0
channels-redis==2.4.0
chardet==3.0.4
Click==7.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
daphne==2.3.0
defusedxml==0.5.0
Django==2.2
django-allauth==0.39.1
django-cors-headers==2.5.2
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
django-storages==1.7.1
djangorestframework==3.9.2
djangorestframework-jwt==1.11.0
djangorestframework-stubs==0.4.2
docutils==0.14
Flask==1.0.2
gevent==1.4.0
greenlet==0.4.15
hiredis==1.0.0
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
itsdangerous==1.1.0
itypes==1.1.0
Jinja2==2.10.1
jmespath==0.9.4
kombu==4.5.0
locust==0.0
locustio==0.11.0
MarkupSafe==1.1.1
msgpack==0.6.1
oauthlib==3.0.1
openapi-codec==1.3.2
psycopg2==2.8.1
PyHamcrest==1.9.0
PyJWT==1.7.1
python-dateutil==2.8.0
python3-openid==3.1.0
pytz==2019.1
PyYAML==5.1
pyzmq==18.0.1
requests==2.21.0
requests-oauthlib==1.2.0
s3transfer==0.2.0
sentry-sdk==0.7.10
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
Twisted==19.2.0
txaio==18.8.1
uritemplate==3.0.0
urllib3==1.24.1
vine==1.3.0
Werkzeug==0.15.2
zope.interface==4.6.0
